requirements.txt
einops==0.8.2
matplotlib==3.10.8
numpy==2.4.2
Pillow==12.1.1
skimage==0.0
torch==2.5.1
torchvision==0.20.1
